requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
streamlit
requests
python-dotenv
openpyxl
xlrd
ydata-profiling
streamlit-aggrid